modularize file writing descriptor creation
