+++ write-file.js
@@ -20,17 +20,6 @@
 	var resolvePath = path.resolve;
 	var getParent = path.dirname;
 
-	const DEFAULT_OPTIONS = freeze({
-		'encoding': 'utf8',
-		'__proto__': null
-	});
-
-	const DEFAULT_DESCRIPTOR = freeze({
-		'data': '',
-		'options': DEFAULT_OPTIONS,
-		'__proto__': null
-	});
-
 	var _writeFile = (filename, descriptor, onfinish, onaction) => {
 		var callOnFinish = (...action) =>
 			onfinish(null, new Info('mkfile', filename, action));
@@ -69,28 +58,6 @@
 			})
 		}, onaction);
 	}
-
-	var _getdesc = (descriptor) => {
-		if (!descriptor) {
-			return DEFAULT_DESCRIPTOR;
-		}
-		switch (typeof descriptor) {
-			case 'object':
-				return {
-					'data': descriptor.data || '',
-					'options': descriptor.options || DEFAULT_OPTIONS,
-					'__proto__': null
-				};
-			case 'string':
-			case 'number':
-				return {
-					'data': descriptor,
-					'options': DEFAULT_OPTIONS,
-					'__proto__': null
-				};
-		}
-		throw new TypeError(`Not supported descriptor format`);
-	};
 
 	module.exports = (filename, descriptor, onfinish, onaction) =>
 		_writeFile(resolvePath(filename), _getdesc(descriptor), _getfunc(onfinish, _throwif), _getfunc(onaction, _donothing));
